requirements.txt
mutagen
requests
pygame
pyyaml
ruamel.yaml
flask
flask-wtf
requests